+++ packages/express-backend/package.json
@@ -2,11 +2,11 @@
   "name": "express-backend",
   "version": "1.0.0",
   "type": "module",
-  "main": "src/server.js",
+  "main": "backend.js",
   "scripts": {
     "test": "dotenv -e .env.test -- jest",
-    "dev": "npx nodemon src/server.js",
-    "start": "node src/server.js"
+    "dev": "npx nodemon backend.js",
+    "start": "node backend.js"
   },
   "author": "",
   "license": "ISC",
